Remove Tailwinds dependency
Don't use Tailwinds anymore. Involves a lot of changes with regards to elements misplaced because of CSS global scoping. The blog is the most heavily affected page - text now doesn't hug images as close as I'd like them to (it looks a little too neat now), but that's okay. Don't want Tailwinds anymore.