Found culprit for wrong Apple touch icon.
The `index.html` file is hidden in the `public` directory. Within this file, we see React auto-generated a link for the Apple touch icon to the be the default React logo. Removed that line and linked my own icon!

Side note: thought of this because was wondering where the React logo was being linked in the first place, since clearly Safari was picking up the default React icon from somewhere within the repo. Did some investigative work and landed here.